pyproject.toml
[project]
name = "BlurtBot"
version = "0.1.0"
description = "A simple discord bot, built for a boot.dev personal project, which does a few rather useless things"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "discord.py==2.6.3",
  "python-dotenv==1.1.0"
]
